Select to be implemeted

+++ BinarySearchTreeExcercise/BinarySearchTree/BinarySearchTree.cs
@@ -170,6 +170,10 @@ public class BinarySearchTree<T> : IBinarySearchTree<T>
 
     public void DeleteMax()
     {
+        if (this.root==null)
+        {
+            throw new InvalidOperationException();
+        }
         var toDelete=this.DeleteMax(this.root);
     }
 
@@ -203,7 +207,27 @@ public class BinarySearchTree<T> : IBinarySearchTree<T>
 
     public int Rank(T element)
     {
-        throw new NotImplementedException();
+        return this.Rank(this.root, element);
+    }
+
+    private int Rank(Node node, T element)
+    {
+        if (node==null)
+        {
+            return 0;
+        }
+
+        int compare = node.Value.CompareTo(element);
+        if (compare>0)
+        {
+            return this.Rank(node.Left, element);
+        }
+        else if(compare<0)
+        {
+            return 1+ this.Count(node.Left)+ this.Rank(node.Right, element);
+        }
+
+        return this.Count(node.Left);
     }
 
     public T Select(int rank)
